requirements.txt
numpy 
pandas 
seaborn 
matplotlib 
scikit-learn 
imblearn
catboost
xgboost
miceforest
kneed
prettytable
pymongo
python-dotenv
pymongo[srv]==4.2.0
-e . 